requirements.txt
streamlit==1.10.0
openai==0.27.0
python-docx==0.8.11
fpdf2==2.5.7
python-dotenv==0.19.2
beautifulsoup4==4.10.0



